Build speedup (#42399)
