+++ src/cata_utility.cpp
@@ -16,6 +16,7 @@
 #include "debug.h"
 #include "filesystem.h"
 #include "json.h"
+#include "ofstream_wrapper.h"
 #include "options.h"
 #include "output.h"
 #include "rng.h"
