+++ src/character_id.cpp
@@ -1,0 +1,8 @@
+#include "character_id.h"
+
+#include <ostream>
+
+std::ostream &operator<<( std::ostream &o, character_id id )
+{
+    return o << id.get_value();
+}
