+++ src/character_id.h
@@ -3,7 +3,7 @@
 #define CATA_SRC_CHARACTER_ID_H
 
 #include <cassert>
-#include <ostream>
+#include <iosfwd>
 
 class JsonIn;
 class JsonOut;
@@ -32,23 +32,25 @@ class character_id
         void serialize( JsonOut & ) const;
         void deserialize( JsonIn & );
 
-        friend inline bool operator==( character_id l, character_id r ) {
-            return l.get_value() == r.get_value();
-        }
-
-        friend inline bool operator!=( character_id l, character_id r ) {
-            return l.get_value() != r.get_value();
-        }
-
-        friend inline bool operator<( character_id l, character_id r ) {
-            return l.get_value() < r.get_value();
-        }
-
-        friend inline std::ostream &operator<<( std::ostream &o, character_id id ) {
-            return o << id.get_value();
-        }
     private:
         int value;
 };
+
+inline bool operator==( character_id l, character_id r )
+{
+    return l.get_value() == r.get_value();
+}
+
+inline bool operator!=( character_id l, character_id r )
+{
+    return l.get_value() != r.get_value();
+}
+
+inline bool operator<( character_id l, character_id r )
+{
+    return l.get_value() < r.get_value();
+}
+
+std::ostream &operator<<( std::ostream &o, character_id id );
 
 #endif // CATA_SRC_CHARACTER_ID_H
