+++ src/mapsharing.cpp
@@ -6,6 +6,7 @@
 
 #include "cata_utility.h"
 #include "filesystem.h"
+#include "ofstream_wrapper.h"
 
 #if defined(__linux__)
 #include <unistd.h>
